Fixed background, spacing, footer, progress on update, save, delete

+++ web/destroy_user.php
@@ -5,13 +5,33 @@ $id = intval($_REQUEST['id']);
 require 'config.ini.php';
 include 'conn.php';
 
-$table = $config['PATTI_USERS_TABLE'];
+$dbase = $config['PATTI_DATABASE'];
+if( $conn = connect($dbase) )
+	{
+		$table = $config['PATTI_USERS_TABLE'];
 
-$sql = "delete from $table where id=$id";
-$result =mysqli_query( $conn, $sql );
-if ($result){
-	echo json_encode(array('success'=>true));
-} else {
+	$sql = "delete from $table where id=$id";
+	if (!$res = $conn->query($sql))
+	{
+						 $err
+					= "QUERY FAIL: "
+					. $sql
+					. ' ERRNO: '
+					. $conn->errno
+					. ' ERROR: '
+					. $conn->error
+					;
+					mysqli_close($conn);
+					echo json_encode(array('errorMsg'=>'Some errors occured.'));
+					return;
+			//		trigger_error($err, E_USER_ERROR);
+	 } // end of got an error
+	 else
+	 {
+			mysqli_close($conn);
+	  	echo json_encode(array('success'=>true));
+		} else {
 	echo json_encode(array('errorMsg'=>'Some errors occured.'));
+}
 }
 ?>
